+++ packages/react-universal/src/createRouterComponent.tsx
@@ -1,7 +1,7 @@
 import React from "react";
 import { RouterProvider, ResponseProvider } from "./Context";
 
-import { CuriRouter, Emitted } from "@curi/types";
+import { CuriRouter, ResponseAndNav } from "@curi/types";
 
 export interface RouterProps {
   children: React.ReactNode;
@@ -11,23 +11,20 @@ export default function createRouterComponent(
   router: CuriRouter
 ): React.FunctionComponent<RouterProps> {
   function initialState() {
-    const { response, navigation } = router.current();
-    return {
-      router,
-      response,
-      navigation
-    };
+    return router.current();
   }
 
   return function Router(props: RouterProps) {
-    const [response, setResponse] = React.useState<Emitted>(initialState);
+    const [response, setResponse] = React.useState<ResponseAndNav>(
+      initialState
+    );
 
     React.useEffect(() => {
       let removed = false;
       const stop = router.observe(
-        (emitted: Emitted) => {
+        ({ response, navigation }) => {
           if (!removed) {
-            setResponse(emitted);
+            setResponse({ response, navigation });
           }
         },
         { initial: false }