+++ packages/react-universal/src/hooks/useResponse.ts
@@ -1,8 +1,8 @@
 import React from "react";
 import { responseContext } from "../Context";
 
-import { Emitted } from "@curi/types";
+import { ResponseAndNav } from "@curi/types";
 
 export default function useResponse() {
-  return React.useContext<Emitted>(responseContext);
+  return React.useContext<ResponseAndNav>(responseContext);
 }